customer-support-openenv/requirements.txt
pydantic>=2.0
openai>=1.0
python-dotenv
fastapi
uvicorn
